# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ChatMessageRole = typeof ChatMessageRole[keyof typeof ChatMessageRole];


export const ChatMessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface ChatMessage {
  role: ChatMessageRole;
  content: string;
}

export interface BotConfig {
  bizName: string;
  bizType?: string;
  /** @nullable */
  services?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  howToOrder?: string | null;
  /** @nullable */
  personality?: string | null;
  /** @nullable */
  welcomeMsg?: string | null;
}

export interface ChatMessageInput {
  messages: ChatMessage[];
  config: BotConfig;
}

export interface ChatResponse {
  content: string;
}

export interface ChatError {
  error: string;
}