# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: chat
    description: Business chatbot operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /chat/send:
    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send a message and get an AI response
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageInput"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatError"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatError"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ChatMessage:
      type: object
      required: [role, content]
      properties:
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
    BotConfig:
      type: object
      required: [bizName]
      properties:
        bizName:
          type: string
        bizType:
          type: string
        services:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        howToOrder:
          type: ["string", "null"]
        personality:
          type: ["string", "null"]
        welcomeMsg:
          type: ["string", "null"]
    ChatMessageInput:
      type: object
      required: [messages, config]
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/ChatMessage"
        config:
          $ref: "#/components/schemas/BotConfig"
    ChatResponse:
      type: object
      required: [content]
      properties:
        content:
          type: string
    ChatError:
      type: object
      required: [error]
      properties:
        error:
          type: string