# lib/api-zod/src/generated/types/botConfig.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BotConfig {
  bizName: string;
  bizType?: string;
  /** @nullable */
  services?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  howToOrder?: string | null;
  /** @nullable */
  personality?: string | null;
  /** @nullable */
  welcomeMsg?: string | null;
}