# lib/api-zod/src/generated/types/chatMessageInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BotConfig } from './botConfig';
import type { ChatMessage } from './chatMessage';

export interface ChatMessageInput {
  messages: ChatMessage[];
  config: BotConfig;
}